+++ RELEASE_NOTES.md
@@ -1,1 +1,4 @@
-Release note for version 5.0.7 not found.
+Improvement release
+
+**Bugfixes**
+- Security update phpunit